refactor: update concurrency pool folder
