+++ packages/server/api/src/app/database/database-connection.ts
@@ -16,7 +16,7 @@ import { OtpEntity } from '../ee/authentication/otp/otp-entity'
 import { ConnectionKeyEntity } from '../ee/connection-keys/connection-key.entity'
 import { CustomDomainEntity } from '../ee/custom-domains/custom-domain.entity'
 import { OAuthAppEntity } from '../ee/oauth-apps/oauth-app.entity'
-import { ConcurrencyPoolEntity } from '../ee/platform/platform-plan/concurrency-pool.entity'
+import { ConcurrencyPoolEntity } from '../ee/platform/concurrency-pool/concurrency-pool.entity'
 import { PlatformPlanEntity } from '../ee/platform/platform-plan/platform-plan.entity'
 import { ProjectMemberEntity } from '../ee/projects/project-members/project-member.entity'
 import { ProjectPlanEntity } from '../ee/projects/project-plan/project-plan.entity'
