+++ packages/server/api/src/app/ee/platform/concurrency-pool/concurrency-pool.entity.ts
@@ -1,8 +1,12 @@
-import { ConcurrencyPool } from '@activepieces/shared'
+import { ConcurrencyPool, Project } from '@activepieces/shared'
 import { EntitySchema } from 'typeorm'
 import { ApIdSchema, BaseColumnSchemaPart } from '../../../database/database-common'
 
-export const ConcurrencyPoolEntity = new EntitySchema<ConcurrencyPool>({
+export type ConcurrencyPoolEntitySchema = ConcurrencyPool & {
+    projects: Project[]
+}
+
+export const ConcurrencyPoolEntity = new EntitySchema<ConcurrencyPoolEntitySchema>({
     name: 'concurrency_pool',
     columns: {
         ...BaseColumnSchemaPart,
@@ -12,4 +16,11 @@ export const ConcurrencyPoolEntity = new EntitySchema<ConcurrencyPool>({
     indices: [
         { name: 'idx_concurrency_pool_platform_id', columns: ['platformId'] },
     ],
+    relations: {
+        projects: {
+            type: 'one-to-many',
+            target: 'project',
+            inverseSide: 'pool',
+        },
+    },
 })
