+++ packages/server/api/src/app/platform/platform.controller.ts
@@ -18,7 +18,7 @@ import { StatusCodes } from 'http-status-codes'
 import { z } from 'zod'
 import { securityAccess } from '../core/security/authorization/fastify-security'
 import { platformToEditMustBeOwnedByCurrentUser } from '../ee/authentication/ee-authorization'
-import { concurrencyPoolService } from '../ee/platform/platform-plan/concurrency-pool.service'
+import { concurrencyPoolService } from '../ee/platform/concurrency-pool/concurrency-pool.service'
 import { platformPlanService } from '../ee/platform/platform-plan/platform-plan.service'
 import { stripeHelper } from '../ee/platform/platform-plan/stripe-helper'
 import { platformProjectService } from '../ee/projects/platform-project-service'
